ChainDetailActionButtons moved to a component
