+++ VultisigApp/VultisigApp/Views/Vault/TokenDetailView.swift
@@ -1,0 +1,18 @@
+//
+//  TokenDetailView.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-05-22.
+//
+
+import SwiftUI
+
+struct TokenDetailView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    TokenDetailView()
+}
